feat: BoxOffice 타입, BoxOfficeResult 타입 구현 및 타입 계층 정리
